requirements.txt
fastapi
uvicorn[standard]
requests
openai
sqlalchemy
alembic
pydantic
python-dotenv
newspaper3k
beautifulsoup4
httpx
python-dateutil
